Add files via upload
In this Scrap we we are collecting all date related with the NBA players in 20/21 season. For this scrap I used the libraries BeautifulSoup and Selenium. With the first library I did web scraping and with the second one I prepared the web page with all data I wanted to collect.